v31.4 Bugfixes and additions

+++ src/TMBuffer.h
@@ -12,7 +12,7 @@
 typedef struct s_tmbuffer
 {
     char Title[32];
-    s32 sx, sy;         // Cursor x and y position
+    s16 sx, sy;         // Cursor x and y position
     s16 HScroll;        // VDP horizontal scroll position
     s16 VScroll;        // VDP vertical scroll position
     u8 ColorFG;
@@ -33,12 +33,12 @@ void TMB_ClearLine(u16 y, u16 line_count);
 void TMB_SetColorFG(u8 v);
 void TMB_MoveCursor(u8 dir, u8 num);
 
-void TMB_SetSX(s32 x);
-s32 TMB_GetSX();
-void TMB_SetSY_A(s32 y);
-s32 TMB_GetSY_A();
-void TMB_SetSY(s32 y);
-s32 TMB_GetSY();
+void TMB_SetSX(s16 x);
+s16 TMB_GetSX();
+void TMB_SetSY_A(s16 y);
+s16 TMB_GetSY_A();
+void TMB_SetSY(s16 y);
+s16 TMB_GetSY();
 
 void TMB_SetVScroll(s16 v);
 void TMB_SetHScroll(s16 h);
